FEATURE: Remove --branch option and replace with optional 'ref' parameter
This change lets a user specify VCS specific object references instead of limiting deployment to the head of a branch. Branches  are valid refs in Git, so superficially the only change in the beam user interface is where a branch is specified on the  command line. Using refs however, allows (at least with Git) deploying a specific commits, tag, or relative object like master~4.