FEATURE: Determine transfer method by config
This update removes the separate rsync, ftp and sftp commands, replacing
them with `up` and `down` commands that make use of the `type` config
key for servers.

[Type]Command classes that previously extended BeamCommand are now named
[Type]TransferMethod and extend TransferMethod. TransferMethods facilitate
creating a deployment provider of the correct type and specifying
command-line options specific to their deployment provider. The options
of a TransferMethod are applied to a DirectionCommand before it runs,
so additional options are seamlessly available based on beam.json and
the contents of argv.

Resolves #13