Get deployer identity using `git config`
Previously, the person who ran `beam up` was identified in the beamlog
by their local username. This isn't ideal, and provides not useful
information if beam is run as root (which I've seen in beamlog files
a few times). This change leverages the information Git already has
to identify a user; usually a full name and email address. Thus, the
beamlog 'Deployer' line changes from:

    Deployer: stephen

to:

    Deployer: Stephen Holdaway <stephen@heyday.co.nz>

If the `git config --get` calls fail, the deployer name is taken from
get_current_user() as before.
.