Fix bug using --no-checksums option with rsync
Files only marked with 'T' in Rsync's itemized changes output were being
added to the beam's list of changes with an empty reason:

    [sent] [] .beamlog

This was happening because running beam with '--no-checksums' was not
falling back to the rsync modified times. With neither comparison method
set, I understand rsync will want to update the modified time of all
files. From the Rsync 3.0.9 man page:

> A t means the modification time is different and is being updated to the
> sender’s value (requires --times).  An alternate value of T means that
> the modification time will be set to the transfer time, which happens
> when a file/symlink/device is updated without  --times and  when  a
> symlink  is changed  and  the  receiver  can’t  set its time. (Note:
> when using an rsync 3.0.0 client, you might see the s flag combined with
> t instead of the proper T flag for this time-setting failure.)