Ignore herzult/php-ssh requirement for ext-ssh2
The PHP ssh2 extension is not required to use beam, however the
herzult/php-ssh package marks it as required. This dependency causes
`composer install` to fail if the extension is not installed. The
SftpDeploymentHandler checks for this extension itself before running,
so this additional enforcement by composer is not required.