Correct behaviour when running with --no-checksums
I introduced an incorrect behaviour in commit e098b909 around 'T' in
rsync's itemized changes output that caused all transfered files to list
'time' as one of the transfer reasons.

Beam shouldn't care about 'T' as it does not indicate a changed file.
Additionally, as beam doesn't have rsync sync mod. times, the use of
timestamps when using --no-checksums is unreliable as the files on the
target will have a modified time of their last deployment.