Allow merging multiple config files at run-time
This change removes the concept of 'applications' (groups of excludes)
and replaces it with the ability to import other beam.json schema configs
using the "import" array in a beam.json file. This allows common config
setups to be used in multiple beam configs without duplication.

Imported configs support all options of a regular beam config, though
imports in imported configs are intentionally ignored. It was prefered
to specify all imports in the main beam config instead of importing
recursively. Eg. instead of:

    [beam.json] <-- imports --< [other-conf.json] <-- which imports -- < [something-else.json]

You would do

    [beam.json] <-- imports --< [other-conf.json]
                <-- imports --< [something-else.json]

This is to allow for greater control over the order of imports.

Application definitions were previously hard-coded in BeamConfiguration,
which made it impossible to add/edit application definitions without
modifying the source of Beam.

Resolves #17