Make remote command limitation more accurate
Previously, if a Deployment provider included LIMITATION_REMOTECOMMAND
as a limitation in `getLimitations`, any defined remote commands would
cause beam to abort, even if the command would not be run (ie. a tagged
command when the tag was not passed). This commit checks if a remote
command will run, and only aborts in the case one would.