Remove unsetting of items in cached FTP directory listings
The unsetting of file entries in Ftp::listCache creates an inconsistency
where the first call with a path will be correct, but subsequent calls
with the same path will return false.

Since Ftp::mkdir() calls Ftp::exists() on each component of a path to
determine what directories should be created, the cache deletion behaviour
in Ftp::exists() could cause mkdir to create a directory that already
existed. This would cause an exception which would abort the deployment.