Try to remove directories using rm on *nix
The rm command is roughly twice as fast as the existing PHP removal
method on my system. For medium size projects, this saves ~150ms.

Note: Some unit tests were failing because the rely on the removeDirectory
helper function which was trying to use `rm` with a vfsStream:// url.
Protocol strings are excluded from running `rm` for this reason.