Add support for sshpass with rsync deployment and ssh commands
Previously when SSH password authentication was used, the user would
be prompted for a password on each new connection to the target. As Beam
usually makes multiple separate SSH connections, this was bothersome in
cases where using key-based authentication wasn’t possible or appropriate.

With this update, the `sshpass` option can be set to true on rsync targets,
and Beam will prompt only once for a password. It is required to have the
sshpass program accessible on PATH for this feature to work.

http://sourceforge.net/projects/sshpass/

It may be possible to achieve this functionality without the use of an
external program, but I couldn’t see an obvious way to do so at this time.

Resolves #39