Allow ability to provide arbitrary command arguments to rsync
Instead of creating options for each rsync feature, this commit allows
the arbitrary use of an —-args="" parameter when using the rsync
deployment provider.

In future additional options may still be provided through the use of an
explicit beam option (via Symfony commands) but this change will
allow the flexibility without needing to foresee what options will be
desirable.