Add force flag to skip checking remote server.
When deploying a large project (~10k files) the initial task of building the file list to deploy (detecting exists()) can take ~30 minutes with the --full flag. Most of that time is in checking remote files exist, by forcibly skipping that step (as I know the remote server is empty) I can cut it down it to ~20 seconds.