Respect command-line options for config interpolations (--ref)
This refactors #48 to use internal state to inform values, rather than
directly calling Git commands when initialising the config. This means
that the deployment context (eg. --ref) is correctly applied to the config
value replacements.

Additionally, the git commands run to generate replacement values are now
run at most one time, and are only run if value replacement actually needs
to be made for a token.