Fix deletion occuring during transfer with delay-updates
The --delay-updates option in Rsync delays updates of files, however deletions
default to the "delete during" mode unless the "--delete-delay" option is given.
The lack of the delayed deletes mode currently is an oversight that means a
target filesystem can be in a partially changed state during a deployment, and
an aborted/parital deployment can cause the target filesystem to lose files
silently.

This issue hasn't caused any noticable issues in over two years of production
use, but it's a potential risk so needed fixing. Picked up while searching the
rsync man page for unrelated reasons.

> The --delete option may be combined with one of the --delete-WHEN options  without  conflict,
> as  well  as --delete-excluded.  However, if none of the --delete-WHEN options are specified,
> rsync will choose the --delete-during algorithm when talking to rsync 3.0.0 or newer, and the
> --delete-before  algorithm  when  talking  to  an  older  rsync.