Fix --help not showing transfer method options since Symfony Console 2.3
Since Symfony Console 2.3, the grouped options we used to show for --help
stopped working. This has been broken for at least a year or two in the
built phar release.

This is a half-fix that at least includes the current target's options when
running `beam up [valid-target] --help`. I'd like to get the grouped options
back in, but it looks like we'll need to implement our own help command to
have control over that formatting/output now.