Add a "syncPermissions" config option for Rsync
It's not always desirable to sync permissions. We've come across this quite
a few times and worked around with --args='--no-p', but it looks like this
will be useful to have in the config.

I feel like this should be off by default, but it's defaulting to true for
now to keep the existing behaviour.

Here are a few examples of when syncing permissions doesn't make sense:

- Permissions on the root directory differ between source and target and the
  target directory is not owned by the ssh user -> permissions error

- Permissions disabled in Git (git config core.fileMode false). The permissions
  from `git archive` are inherited from the source system and aren't meaningful.

- The target system has a different umask to the source and this needs to be
  honoured. For example, some shared hosting will refuse to serve PHP files
  that have group write permission, which currently requires a Beam command
  to chmod before deploying.

- Custom acls (eg. setfacl) or some other permissions scheme is in use on the
  target that specifies how permissions are inherited.

Resolves #53